refactor: extract shared styles and page helpers
