+++ src/pages/LoginPage/index.tsx
@@ -10,7 +10,6 @@ import Snackbar from '@mui/material/Snackbar'
 import Stack from '@mui/material/Stack'
 import TextField from '@mui/material/TextField'
 import Typography from '@mui/material/Typography'
-import { FirebaseError } from 'firebase/app'
 
 import Button from '@/components/Button'
 import {
@@ -19,72 +18,9 @@ import {
   loginWithEmail,
   resendVerificationEmail,
 } from '@/firebase/auth'
+import { centeredPageSx, formFieldSx, formPageSx } from '@/styles/page'
 
-const pageSx = {
-  minHeight: 'calc(100vh - 91px)',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-  px: 3,
-  py: { xs: 5, md: 7 },
-}
-
-const formSx = {
-  width: '100%',
-  maxWidth: 520,
-}
-
-const fieldSx = {
-  '& .MuiOutlinedInput-root': {
-    backgroundColor: 'background.paper',
-  },
-}
-
-function getLoginErrorMessage(error: unknown) {
-  if (error instanceof FirebaseError) {
-    switch (error.code) {
-      case 'auth/invalid-email':
-        return 'Email: enter a valid email address.'
-      case 'auth/missing-password':
-        return 'Password: enter your password.'
-      case 'auth/invalid-credential':
-        return 'Email or password is incorrect.'
-      case 'auth/popup-closed-by-user':
-        return 'Google sign in was closed before it finished.'
-      case 'auth/popup-blocked':
-        return 'Google sign in popup was blocked by the browser.'
-      case 'auth/account-exists-with-different-credential':
-        return 'This email is already registered with another sign in method.'
-      case 'auth/operation-not-allowed':
-        return 'Firebase Auth: this sign in method is not enabled.'
-      case 'auth/network-request-failed':
-        return 'Network: check your internet connection and try again.'
-      case 'auth/too-many-requests':
-        return 'Too many attempts. Wait a few minutes before trying again.'
-      case 'permission-denied':
-        return 'Firestore: permission denied while saving your visitor profile.'
-      default:
-        return `${error.code}: ${error.message}`
-    }
-  }
-
-  if (error instanceof Error && error.message === EMAIL_NOT_VERIFIED_ERROR) {
-    return 'Email verification required. Verify your email before logging in.'
-  }
-
-  if (
-    error instanceof Error &&
-    error.message === 'verification-email-password-only'
-  ) {
-    return 'Verification email can only be resent for email/password accounts.'
-  }
-
-  if (error instanceof Error && error.message === 'email-already-verified') {
-    return 'This email is already verified. Log in to continue.'
-  }
-
-  return 'Unable to log in. Please try again.'
-}
+import { getLoginErrorMessage } from './utils'
 
 function LoginPage() {
   const navigate = useNavigate()
@@ -161,8 +97,8 @@ function LoginPage() {
 
   return (
     <>
-      <Box component="main" sx={pageSx}>
-        <Stack component="form" onSubmit={handleSubmit} spacing={3} sx={formSx}>
+      <Box component="main" sx={centeredPageSx}>
+        <Stack component="form" onSubmit={handleSubmit} spacing={3} sx={formPageSx}>
           <Box>
             <Typography variant="h3">Log in</Typography>
             <Typography color="text.secondary" sx={{ mt: 1 }} variant="body1">
@@ -192,7 +128,7 @@ function LoginPage() {
             label="Email"
             onChange={(event) => setEmail(event.target.value)}
             required
-            sx={fieldSx}
+            sx={formFieldSx}
             type="email"
             value={email}
           />
@@ -202,7 +138,7 @@ function LoginPage() {
             label="Password"
             onChange={(event) => setPassword(event.target.value)}
             required
-            sx={fieldSx}
+            sx={formFieldSx}
             type="password"
             value={password}
           />
