+++ src/pages/ProfilePage/styles.ts
@@ -1,0 +1,19 @@
+export const profileImageButtonSx = {
+  width: 84,
+  minHeight: 102,
+  flexDirection: 'column',
+  gap: 1,
+  borderRadius: '8px',
+  color: 'text.primary',
+  border: '1px solid',
+  borderColor: 'divider',
+  p: 1,
+  '&:hover': {
+    backgroundColor: 'action.hover',
+  },
+}
+
+export const selectedProfileImageButtonSx = {
+  borderColor: 'primary.main',
+  backgroundColor: 'background.paper',
+}
